delete test class and integrate tests into correct test class already present
