+++ ff-parser/src/test/java/org/uniprot/core/flatfile/parser/integration/UniProtKBEntryIteratorIT.java
@@ -1,11 +1,16 @@
 package org.uniprot.core.flatfile.parser.integration;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.List;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.uniprot.core.flatfile.parser.impl.DefaultUniProtEntryIterator;
@@ -14,17 +19,23 @@ import org.uniprot.core.flatfile.writer.impl.UniProtFlatfileWriter;
 import org.uniprot.core.uniprotkb.UniProtKBEntry;
 
 class UniProtKBEntryIteratorIT {
+    private DefaultUniProtEntryIterator entryIterator;
+
+    @BeforeEach
+    void setUp() {
+        entryIterator = new DefaultUniProtEntryIterator(2, 5, 10);
+        entryIterator.setIgnoreWrong(true);
+    }
+
     @Test
     void testSingleTxl() {
         try {
-            DefaultUniProtEntryIterator iterator = new DefaultUniProtEntryIterator();
-            iterator.setIgnoreWrong(true);
             String filename = "src/test/resources/entryIT/A9N0W4.txl";
-            iterator.setInput(filename, "", "", "", "");
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry = iterator.next();
+            entryIterator.setInput(filename, "", "", "", "");
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry = entryIterator.next();
             assertNotNull(entry);
-            assertFalse(iterator.hasNext());
+            assertFalse(entryIterator.hasNext());
         } catch (Exception e) {
             fail("test failed");
         }
@@ -33,13 +44,11 @@ class UniProtKBEntryIteratorIT {
     @Test
     void testSingleTxlRoundTrip() {
         try {
-            DefaultUniProtEntryIterator iterator = new DefaultUniProtEntryIterator();
-            iterator.setIgnoreWrong(true);
             String filename = "src/test/resources/entryIT/A9N0W4.txl";
             String entryStr = readFile(filename);
-            iterator.setInput(filename, "", "", "", "");
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry = iterator.next();
+            entryIterator.setInput(filename, "", "", "", "");
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry = entryIterator.next();
             FlatfileWriter<UniProtKBEntry> writer = new UniProtFlatfileWriter();
             String convertedEntryStr = writer.write(entry, false);
             assertEquals(entryStr, convertedEntryStr + "\n");
@@ -51,14 +60,12 @@ class UniProtKBEntryIteratorIT {
     @Test
     void testSingleDat() {
         try {
-            DefaultUniProtEntryIterator iterator = new DefaultUniProtEntryIterator();
-            iterator.setIgnoreWrong(true);
             String filename = "src/test/resources/entryIT/Q32K04.dat";
-            iterator.setInput(filename, "", "", "", "");
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry = iterator.next();
+            entryIterator.setInput(filename, "", "", "", "");
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry = entryIterator.next();
             assertNotNull(entry);
-            assertFalse(iterator.hasNext());
+            assertFalse(entryIterator.hasNext());
         } catch (Exception e) {
             fail("test failed");
         }
@@ -67,13 +74,11 @@ class UniProtKBEntryIteratorIT {
     @Test
     void testSingleDatRoundTrip() {
         try {
-            DefaultUniProtEntryIterator iterator = new DefaultUniProtEntryIterator();
-            iterator.setIgnoreWrong(true);
             String filename = "src/test/resources/entryIT/Q32K04.dat";
             String entryStr = readFile(filename);
-            iterator.setInput(filename, "", "", "", "");
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry = iterator.next();
+            entryIterator.setInput(filename, "", "", "", "");
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry = entryIterator.next();
             FlatfileWriter<UniProtKBEntry> writer = new UniProtFlatfileWriter();
             String convertedEntryStr = writer.write(entry, true);
             System.out.println(convertedEntryStr);
@@ -86,18 +91,16 @@ class UniProtKBEntryIteratorIT {
     @Test
     void testMultiDat() {
         try {
-            DefaultUniProtEntryIterator iterator = new DefaultUniProtEntryIterator();
-            iterator.setIgnoreWrong(true);
             String filename = "src/test/resources/entryIT/A8EZU1_D6RDV7.dat";
-            iterator.setInput(filename, "", "", "", "");
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry = iterator.next();
+            entryIterator.setInput(filename, "", "", "", "");
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry = entryIterator.next();
 
             assertNotNull(entry);
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry2 = iterator.next();
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry2 = entryIterator.next();
             assertNotNull(entry2);
-            assertFalse(iterator.hasNext());
+            assertFalse(entryIterator.hasNext());
         } catch (Exception e) {
             fail("test failed");
         }
@@ -106,12 +109,10 @@ class UniProtKBEntryIteratorIT {
     @Disabled
     void testMultiDatRoundTrip() {
         try {
-            DefaultUniProtEntryIterator iterator = new DefaultUniProtEntryIterator();
-            iterator.setIgnoreWrong(true);
             String filename = "src/test/resources/entryIT/A8EZU1_D6RDV7.dat.gz";
-            iterator.setInput(filename, "", "", "", "");
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry = iterator.next();
+            entryIterator.setInput(filename, "", "", "", "");
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry = entryIterator.next();
             assertNotNull(entry);
             String filename1 = "src/test/resources/entryIT/A8EZU1.dat";
             String entryStr1 = readFile(filename1);
@@ -119,14 +120,14 @@ class UniProtKBEntryIteratorIT {
             String convertedEntryStr1 = writer.write(entry, true);
             assertEquals(entryStr1, convertedEntryStr1 + "\n");
 
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry2 = iterator.next();
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry2 = entryIterator.next();
             assertNotNull(entry2);
             String filename2 = "src/test/resources/entryIT/D6RDV7.dat";
             String entryStr2 = readFile(filename2);
             String convertedEntryStr2 = writer.write(entry2, true);
             assertEquals(entryStr2, convertedEntryStr2 + "\n");
-            assertFalse(iterator.hasNext());
+            assertFalse(entryIterator.hasNext());
         } catch (Exception e) {
             fail("test failed");
         }
@@ -135,20 +136,43 @@ class UniProtKBEntryIteratorIT {
     @Test
     void testMultiDatGz() {
         try {
-            DefaultUniProtEntryIterator iterator = new DefaultUniProtEntryIterator();
-            iterator.setIgnoreWrong(true);
             String filename = "src/test/resources/entryIT/A8EZU1_D6RDV7.dat.gz";
-            iterator.setInput(filename, "", "", "", "");
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry = iterator.next();
+            entryIterator.setInput(filename, "", "", "", "");
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry = entryIterator.next();
             assertNotNull(entry);
-            assertTrue(iterator.hasNext());
-            UniProtKBEntry entry2 = iterator.next();
+            assertTrue(entryIterator.hasNext());
+            UniProtKBEntry entry2 = entryIterator.next();
             assertNotNull(entry2);
-            assertFalse(iterator.hasNext());
+            assertFalse(entryIterator.hasNext());
         } catch (Exception e) {
             fail("test failed");
         }
+    }
+
+    @Test
+    void parseErrorForAllEntriesDoesNotCauseHanging() {
+        entryIterator.setInput(
+                "src/test/resources/entryIT/ERROR_ERROR.dat", null, null, null, null);
+        while (entryIterator.hasNext()) {
+            UniProtKBEntry entry =
+                    entryIterator
+                            .next(); // the entry should be null because there was a parse error
+            assertThat(entry, is(nullValue()));
+        }
+    }
+
+    @Test
+    void parseErrorInOnlyOneEntryDoesNotCauseHanging() {
+        entryIterator.setInput(
+                "src/test/resources/entryIT/A8EZU1_ERROR_D6RDV7.dat", null, null, null, null);
+        List<String> accessions = new ArrayList<>();
+        while (entryIterator.hasNext()) {
+            UniProtKBEntry entry = entryIterator.next();
+            accessions.add(entry.getPrimaryAccession().getValue());
+        }
+
+        assertThat(accessions, containsInAnyOrder("A8EZU1", "D6RDV7"));
     }
 
     private String readFile(String file) throws IOException {
